services/slot_manager/requirements.txt
kafka-python
redis
pydantic